refactor: underscore internal-only widget modules
Match the codebase convention (session/_core, otb/_base) for modules with no
importers outside their own subpackage:
- signals/{plot,controls,state,scan} -> _plot/_controls/_state/_scan
- training/session_state -> _session_state
Public entry points (viewer/raw/stream_panel, session_manager) and
cross-subpackage-shared helpers (signals/transforms, panels/log_box) keep their
bare names. Internal cross-imports updated.
