+++ myogestic/widgets/signals/_scan.py
