+++ myogestic/widgets/signals/_state.py
