+++ myogestic/widgets/training/_session_state.py
