+++ myogestic/widgets/training/session_manager.py
@@ -8,7 +8,7 @@ from imgui_bundle import portable_file_dialogs as pfd
 
 from myogestic.contracts import TrainingData
 from myogestic.widgets.common import PALETTE, panel_header
-from myogestic.widgets.training.session_state import (
+from myogestic.widgets.training._session_state import (
     SessionWidgetState,
     add_recorded_session,
     class_pool_and_active,
